Fix memory leaks.
The first obvious change was to free the 'input' string in 'main'. I also
added a 'free_tokens' function that frees the buffers in the 'tokens'
struct. A third, less obvious, leak was in 'add_token'. Instead of
adding the string it is given to 'tokens_list' it copies it. The former
string's memory then leaks. There is no good reason for copying
the string instead of just storing it, so I eliminated the copy.
Incidentally, this was the only place where the 'string_copy' function
was used, so I removed it.